+++ src/Resources/contao/languages/de/tl_fernschach_spieler_mails.php
@@ -10,7 +10,7 @@
  * Backend-Modul Übersetzungen
  */
 
-$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['main']= array('E-Mail', 'E-Mail anlegen');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['templates'] = array('E-Mail-Templates', 'Globale E-Mail-Templates verwalten');
 
 // Standardfunktionen
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['new']= array('Neue E-Mail', 'Neue E-Mail anlegen');
@@ -21,6 +21,9 @@ $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['send']= array("E-Mail %s ver
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['show']= array("Details zur E-Mail %s anzeigen", "Details zur E-Mail %s anzeigen");
 
 // Formularfelder
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['template_legend'] = 'Vorlage';
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['template']= array('Vorlage wählen', 'Wählen Sie eine Vorlage. Bei einer Vorlagenänderung wird automatisch die Vorschau unten neu generiert.');
+
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['text_legend'] = 'Betreff und Text';
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['subject']= array('Betreff', 'Betreffzeile der E-Mail');
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['content']= array('Text', 'Textinhalt der E-Mail');
@@ -28,6 +31,9 @@ $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['content']= array('Text', 'Te
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur_legend'] = 'Signatur';
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur']= array('Signatur aktivieren', 'Signatur für diese E-Mail aktivieren.');
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['signatur_text']= array('Signaturtext', 'Die Signatur kann im Profil voreingestellt werden.');
+
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['preview_legend'] = 'Vorschau';
+$GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['preview']= array('Vorschau', 'Vorschau mit der gewählten Vorlage. Die Vorschau wird neu generiert, wenn Sie den Datensatz speichern oder die Vorlage wechseln.');
 
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['mail_legend'] = 'E-Mail-Optionen';
 $GLOBALS['TL_LANG']['tl_fernschach_spieler_mails']['copyBenutzer']= array('E-Mail-Kopie an mich', 'Eine E-Mail-Kopie wird an mich (aktueller Benutzer) gesendet.');